fixed segfault in cutting the header out of the buffer
